folder handling fix

+++ src/piler.h
@@ -55,8 +55,8 @@ int retrieve_email_from_archive(struct session_data *sdata, struct __data *data,
 int file_from_archive_to_network(char *filename, int sd, int tls_enable, struct __data *data, struct __config *cfg);
 
 int import_message(char *filename, struct session_data *sdata, struct __data *data, struct __config *cfg);
-unsigned long get_folder_id(struct session_data *sdata, struct __data *data, char *foldername, int parent_id);
-unsigned long add_new_folder(struct session_data *sdata, struct __data *data, char *foldername, int parent_id);
+int get_folder_id(struct session_data *sdata, struct __data *data, char *foldername, int parent_id);
+int add_new_folder(struct session_data *sdata, struct __data *data, char *foldername, int parent_id);
 
 int store_index_data(struct session_data *sdata, struct _state *state, struct __data *data, uint64 id, struct __config *cfg);
